use separate registration paths
