+++ pkg/gitprovider/bitbucket/cloud/provider_test.go
@@ -144,6 +144,30 @@ func intPtr(i int) *int       { return &i }
 
 func statePtr(s PullrequestState) *PullrequestState { return &s }
 
+func Test_registration(t *testing.T) {
+	t.Run("predicate matches bitbucket.org", func(t *testing.T) {
+		assert.True(t, registration.Predicate("https://bitbucket.org/owner/repo"))
+	})
+
+	t.Run("predicate does not match self-hosted URLs", func(t *testing.T) {
+		assert.False(t, registration.Predicate("https://bitbucket.example.com/projects/PROJ/repos/repo"))
+	})
+
+	t.Run("predicate does not match other providers", func(t *testing.T) {
+		assert.False(t, registration.Predicate("https://github.com/owner/repo"))
+	})
+
+	t.Run("predicate handles invalid URLs", func(t *testing.T) {
+		assert.False(t, registration.Predicate("://invalid-url"))
+	})
+
+	t.Run("NewProvider factory works", func(t *testing.T) {
+		p, err := registration.NewProvider("https://bitbucket.org/owner/repo", nil)
+		assert.NoError(t, err)
+		assert.NotNil(t, p)
+	})
+}
+
 func TestNewProvider(t *testing.T) {
 	t.Run("successful creation", func(t *testing.T) {
 		provider, err := NewProvider("https://bitbucket.org/owner/repo", &gitprovider.Options{Token: "token"})
